+++ packages/@expo/cli/src/start/server/metro/__tests__/withMetroMultiPlatform.test.ts
@@ -141,6 +141,43 @@ describe(withExtendedResolver, () => {
     );
   });
 
+  it(`resolves a file with module condition for ESM`, async () => {
+    mockMinFs();
+
+    const modified = withExtendedResolver(asMetroConfig({ projectRoot: '/root/' }), {
+      tsconfig: null,
+      isTsconfigPathsEnabled: false,
+      getMetroBundler: getMetroBundlerGetter(),
+    });
+
+    const platform = 'ios';
+
+    modified.resolver.resolveRequest!(
+      {
+        ...getDefaultRequestContext(),
+        isESMImport: true,
+      },
+      'react-native',
+      platform
+    );
+
+    expect(getResolveFunc()).toHaveBeenCalledTimes(1);
+    expect(getResolveFunc()).toHaveBeenCalledWith(
+      expect.objectContaining({
+        extraNodeModules: {},
+        mainFields: ['react-native', 'browser', 'module', 'main'],
+        nodeModulesPaths: ['/root/node_modules'],
+        preferNativePlatform: true,
+        sourceExts: ['mjs', 'ts', 'tsx', 'js', 'jsx', 'json', 'css'],
+        customResolverOptions: {},
+        originModulePath: expect.anything(),
+        getPackage: expect.any(Function),
+      }),
+      'react-native',
+      platform
+    );
+  });
+
   it(`resolves against tsconfig baseUrl`, async () => {
     mockMinFs();
 
@@ -208,7 +245,7 @@ describe(withExtendedResolver, () => {
     expect(getResolveFunc()).toHaveBeenCalledTimes(1);
     expect(getResolveFunc()).toHaveBeenCalledWith(
       expect.objectContaining({
-        mainFields: ['browser', 'module', 'main'],
+        mainFields: ['browser', 'main'],
         preferNativePlatform: false,
       }),
       '/src/react-native',
@@ -231,7 +268,7 @@ describe(withExtendedResolver, () => {
     expect(getResolveFunc()).toHaveBeenCalledTimes(1);
     expect(getResolveFunc()).toHaveBeenCalledWith(
       expect.objectContaining({
-        mainFields: ['browser', 'module', 'main'],
+        mainFields: ['browser', 'main'],
         preferNativePlatform: false,
       }),
       'react-native-web',
@@ -258,7 +295,7 @@ describe(withExtendedResolver, () => {
     expect(getResolveFunc()).toHaveBeenCalledTimes(1);
     expect(getResolveFunc()).toHaveBeenCalledWith(
       expect.objectContaining({
-        mainFields: ['browser', 'module', 'main'],
+        mainFields: ['browser', 'main'],
         preferNativePlatform: false,
       }),
       'expo-asset/build/resolveAssetSource',
@@ -473,7 +510,7 @@ describe(withExtendedResolver, () => {
     expect(getResolveFunc()).toHaveBeenCalledTimes(1);
     expect(getResolveFunc()).toHaveBeenCalledWith(
       expect.objectContaining({
-        mainFields: ['browser', 'module', 'main'],
+        mainFields: ['browser', 'main'],
         preferNativePlatform: false,
       }),
       'node:path',
@@ -510,7 +547,7 @@ describe(withExtendedResolver, () => {
     expect(getResolveFunc()).toHaveBeenCalledWith(
       expect.objectContaining({
         nodeModulesPaths: ['/root/node_modules'],
-        mainFields: ['browser', 'module', 'main'],
+        mainFields: ['browser', 'main'],
         preferNativePlatform: false,
       }),
       'node:path',
@@ -566,6 +603,7 @@ describe(withExtendedResolver, () => {
     modified.resolver.resolveRequest!(
       {
         ...getDefaultRequestContext(),
+        isESMImport: true,
         customResolverOptions: {
           environment: 'react-server',
         },
@@ -589,6 +627,7 @@ describe(withExtendedResolver, () => {
         unstable_conditionsByPlatform: {},
         unstable_enablePackageExports: true,
         getPackage: expect.any(Function),
+        isESMImport: true,
       },
       'react-foobar',
       platform
@@ -605,6 +644,7 @@ describe(withExtendedResolver, () => {
 
     const platform = 'web';
 
+    // CommonJS
     modified.resolver.resolveRequest!(
       {
         ...getDefaultRequestContext(),
@@ -617,7 +657,40 @@ describe(withExtendedResolver, () => {
     );
 
     expect(getResolveFunc()).toHaveBeenCalledTimes(1);
-    expect(getResolveFunc()).toHaveBeenCalledWith(
+    expect(getResolveFunc()).toHaveBeenLastCalledWith(
+      {
+        customResolverOptions: { environment: 'react-server' },
+        dev: true,
+        extraNodeModules: {},
+        mainFields: ['main', 'module'],
+        nodeModulesPaths: ['/root/node_modules'],
+        originModulePath: '/root/index.js',
+        preferNativePlatform: false,
+        sourceExts: ['ts', 'tsx', 'js', 'jsx', 'mjs', 'json', 'css'],
+        unstable_conditionNames: ['node', 'react-server', 'workerd'],
+        unstable_conditionsByPlatform: {},
+        unstable_enablePackageExports: true,
+        getPackage: expect.any(Function),
+      },
+      'react-foobar',
+      platform
+    );
+
+    // ESM
+    modified.resolver.resolveRequest!(
+      {
+        ...getDefaultRequestContext(),
+        isESMImport: true,
+        customResolverOptions: {
+          environment: 'react-server',
+        },
+      },
+      'react-foobar',
+      platform
+    );
+
+    expect(getResolveFunc()).toHaveBeenCalledTimes(2);
+    expect(getResolveFunc()).toHaveBeenLastCalledWith(
       {
         customResolverOptions: { environment: 'react-server' },
         dev: true,
@@ -631,6 +704,7 @@ describe(withExtendedResolver, () => {
         unstable_conditionsByPlatform: {},
         unstable_enablePackageExports: true,
         getPackage: expect.any(Function),
+        isESMImport: true,
       },
       'react-foobar',
       platform